Fix word generator taking empty words
Previous fix we had merely ignored the last line from word list (which
is commonly is some kind of whitespace, usually newline). But for word
lists generated with Clojure scripts, it effectively ignored the last
word, which is not ideal.

This fixes that by checking whether the word we are pulling in is not
empty